# gpt-actions.openapi.yaml
openapi: 3.1.0
info:
  title: Windows LocalControl GPT Actions
  description: Private Custom GPT action schema for a single-user Windows LocalControl
    bridge. Full-control execution mode is enabled; use /execution/logs to retrieve
    command and terminal output.
  version: 0.1.0
paths:
  /fs/list:
    post:
      tags:
      - filesystem
      summary: Fs List
      operationId: fs_list
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FsListRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FsListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /fs/read:
    post:
      tags:
      - filesystem
      summary: Fs Read
      operationId: fs_read
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FsReadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FsReadResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /fs/write:
    post:
      tags:
      - filesystem
      summary: Fs Write
      operationId: fs_write
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FsWriteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FsWriteResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /fs/replace:
    post:
      tags:
      - filesystem
      summary: Fs Replace
      operationId: fs_replace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FsReplaceRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FsReplaceResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /fs/delete:
    post:
      tags:
      - filesystem
      summary: Fs Delete
      operationId: fs_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FsDeleteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FsDeleteResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /fs/stat:
    post:
      tags:
      - filesystem
      summary: Fs Stat
      operationId: fs_stat
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FsStatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FsStatResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /search/files:
    post:
      tags:
      - search
      summary: Search Files Endpoint
      operationId: search_files
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchFilesRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchFilesResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /search/content:
    post:
      tags:
      - search
      summary: Search Content Endpoint
      operationId: search_content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchContentRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchContentResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /projects/register:
    post:
      tags:
      - projects
      summary: Project Register
      operationId: project_register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRegisterRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /projects/list:
    post:
      tags:
      - projects
      summary: Project List
      operationId: project_list
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /system/info:
    get:
      tags:
      - system
      summary: System Info Endpoint
      operationId: system_info
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInfoResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
components:
  schemas:
...
    ContentMatch:
      properties:
        path:
          type: string
          title: Path
        line_number:
          type: integer
          title: Line Number
        line:
          type: string
          title: Line
        redactions:
          type: integer
          title: Redactions
      additionalProperties: false
      type: object
      required:
      - path
      - line_number
      - line
      - redactions
      title: ContentMatch
    ErrorResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: false
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
      additionalProperties: false
      type: object
      required:
      - code
      - message
      title: ErrorResponse
...
    FileInfo:
      properties:
        path:
          type: string
          title: Path
        name:
          type: string
          title: Name
        is_dir:
          type: boolean
          title: Is Dir
        size:
          anyOf:
          - type: integer
          - type: 'null'
          title: Size
        modified_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Modified At
      additionalProperties: false
      type: object
      required:
      - path
      - name
      - is_dir
      title: FileInfo
    FsDeleteRequest:
      properties:
        path:
          type: string
          title: Path
        recursive:
          type: boolean
          title: Recursive
          default: false
      additionalProperties: false
      type: object
      required:
      - path
      title: FsDeleteRequest
    FsDeleteResponse:
      properties:
        path:
          type: string
          title: Path
        deleted:
          type: boolean
          title: Deleted
      additionalProperties: false
      type: object
      required:
      - path
      - deleted
      title: FsDeleteResponse
    FsListRequest:
      properties:
        path:
          type: string
          title: Path
        recursive:
          type: boolean
          title: Recursive
          default: false
        include_hidden:
          type: boolean
          title: Include Hidden
          default: false
        max_entries:
          type: integer
          maximum: 5000.0
          minimum: 1.0
          title: Max Entries
          default: 200
      additionalProperties: false
      type: object
      required:
      - path
      title: FsListRequest
    FsListResponse:
      properties:
        path:
          type: string
          title: Path
        entries:
          items:
            $ref: '#/components/schemas/FileInfo'
          type: array
          title: Entries
        truncated:
          type: boolean
          title: Truncated
      additionalProperties: false
      type: object
      required:
      - path
      - entries
      - truncated
      title: FsListResponse
    FsReadRequest:
      properties:
        path:
          type: string
          title: Path
        encoding:
          type: string
          title: Encoding
          default: utf-8
        max_bytes:
          type: integer
          maximum: 10485760.0
          minimum: 1.0
          title: Max Bytes
          default: 65536
        include_secrets:
          type: boolean
          title: Include Secrets
          default: false
      additionalProperties: false
      type: object
      required:
      - path
      title: FsReadRequest
    FsReadResponse:
      properties:
        path:
          type: string
          title: Path
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
        content_base64:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Base64
        encoding:
          type: string
          title: Encoding
        binary:
          type: boolean
          title: Binary
        bytes_read:
          type: integer
          title: Bytes Read
        truncated:
          type: boolean
          title: Truncated
        redactions:
          type: integer
          title: Redactions
      additionalProperties: false
      type: object
      required:
      - path
      - encoding
      - binary
      - bytes_read
      - truncated
      - redactions
      title: FsReadResponse
    FsReplaceRequest:
      properties:
        path:
          type: string
          title: Path
        old:
          type: string
          title: Old
        new:
          type: string
          title: New
        count:
          type: integer
          minimum: 0.0
          title: Count
          description: 0 means replace every occurrence.
          default: 0
        encoding:
          type: string
          title: Encoding
          default: utf-8
        create_backup:
          type: boolean
          title: Create Backup
          default: true
      additionalProperties: false
      type: object
      required:
      - path
      - old
      - new
      title: FsReplaceRequest
    FsReplaceResponse:
      properties:
        path:
          type: string
          title: Path
        replacements:
          type: integer
          title: Replacements
        backup_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Backup Path
        file:
          $ref: '#/components/schemas/FileInfo'
      additionalProperties: false
      type: object
      required:
      - path
      - replacements
      - backup_path
      - file
      title: FsReplaceResponse
    FsStatRequest:
      properties:
        path:
          type: string
          title: Path
      additionalProperties: false
      type: object
      required:
      - path
      title: FsStatRequest
    FsStatResponse:
      properties:
        file:
          $ref: '#/components/schemas/FileInfo'
        exists:
          type: boolean
          title: Exists
      additionalProperties: false
      type: object
      required:
      - file
      - exists
      title: FsStatResponse
    FsWriteRequest:
      properties:
        path:
          type: string
          title: Path
        content:
          type: string
          title: Content
        encoding:
          type: string
          title: Encoding
          default: utf-8
        create_parents:
          type: boolean
          title: Create Parents
          default: true
        overwrite:
          type: boolean
          title: Overwrite
          default: false
        append:
          type: boolean
          title: Append
          default: false
      additionalProperties: false
      type: object
      required:
      - path
      - content
      title: FsWriteRequest
    FsWriteResponse:
      properties:
        path:
          type: string
          title: Path
        bytes_written:
          type: integer
          title: Bytes Written
        mode:
          type: string
          enum:
          - created
          - overwritten
          - appended
          title: Mode
        file:
          $ref: '#/components/schemas/FileInfo'
      additionalProperties: false
      type: object
      required:
      - path
      - bytes_written
      - mode
      - file
      title: FsWriteResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProjectInfo:
      properties:
        project_id:
          type: string
          title: Project Id
        name:
          type: string
          title: Name
        path:
          type: string
          title: Path
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          type: string
          title: Created At
        last_used_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Used At
        exists:
          type: boolean
          title: Exists
      additionalProperties: false
      type: object
      required:
      - project_id
      - name
      - path
      - created_at
      - exists
      title: ProjectInfo
    ProjectListResponse:
      properties:
        projects:
          items:
            $ref: '#/components/schemas/ProjectInfo'
          type: array
          title: Projects
      additionalProperties: false
      type: object
      required:
      - projects
      title: ProjectListResponse
    ProjectRegisterRequest:
      properties:
        path:
          type: string
          title: Path
        project_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      additionalProperties: false
      type: object
      required:
      - path
      title: ProjectRegisterRequest
...
    SearchContentRequest:
      properties:
        root:
          anyOf:
          - type: string
          - type: 'null'
          title: Root
        project_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
        pattern:
          type: string
          title: Pattern
        glob:
          anyOf:
          - type: string
          - type: 'null'
          title: Glob
        regex:
          type: boolean
          title: Regex
          default: false
        case_sensitive:
          type: boolean
          title: Case Sensitive
          default: false
        recursive:
          type: boolean
          title: Recursive
          default: true
        include_hidden:
          type: boolean
          title: Include Hidden
          default: false
        max_results:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Max Results
          default: 100
        max_file_bytes:
          type: integer
          maximum: 10485760.0
          minimum: 1024.0
          title: Max File Bytes
          default: 1048576
      additionalProperties: false
      type: object
      required:
      - pattern
      title: SearchContentRequest
    SearchContentResponse:
      properties:
        root:
          type: string
          title: Root
        matches:
          items:
            $ref: '#/components/schemas/ContentMatch'
          type: array
          title: Matches
        truncated:
          type: boolean
          title: Truncated
      additionalProperties: false
      type: object
      required:
      - root
      - matches
      - truncated
      title: SearchContentResponse
    SearchFilesRequest:
      properties:
        root:
          anyOf:
          - type: string
          - type: 'null'
          title: Root
        project_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
        query:
          anyOf:
          - type: string
          - type: 'null'
          title: Query
        glob:
          anyOf:
          - type: string
          - type: 'null'
          title: Glob
        recursive:
          type: boolean
          title: Recursive
          default: true
        include_hidden:
          type: boolean
          title: Include Hidden
          default: false
        max_results:
          type: integer
          maximum: 5000.0
          minimum: 1.0
          title: Max Results
          default: 100
      additionalProperties: false
      type: object
      title: SearchFilesRequest
    SearchFilesResponse:
      properties:
        root:
          type: string
          title: Root
        results:
          items:
            $ref: '#/components/schemas/FileInfo'
          type: array
          title: Results
        truncated:
          type: boolean
          title: Truncated
      additionalProperties: false
      type: object
      required:
      - root
      - results
      - truncated
      title: SearchFilesResponse
...
    SystemInfoResponse:
      properties:
        os:
          type: string
          title: Os
        platform:
          type: string
          title: Platform
        hostname:
          type: string
          title: Hostname
        user:
          type: string
          title: User
        process_id:
          type: integer
          title: Process Id
        python:
          type: string
          title: Python
        cwd:
          type: string
          title: Cwd
        is_admin:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Admin
      additionalProperties: false
      type: object
      required:
      - os
      - platform
      - hostname
      - user
      - process_id
      - python
      - cwd
      - is_admin
      title: SystemInfoResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://YOUR-RESERVED-NGROK-DOMAIN.ngrok-free.app